+++ tweet-service/src/main/java/com/tweetapp/tweetservice/repository/TweetRepository.java
@@ -1,9 +1,9 @@
-package com.fse1.tweetservice.repository;
+package com.tweetapp.tweetservice.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.fse1.tweetservice.entity.TweetEntity;
+import com.tweetapp.tweetservice.entity.TweetEntity;
 
 @Repository
 public interface TweetRepository extends MongoRepository<TweetEntity, String> {
