+++ tweet-service/src/main/java/com/tweetapp/tweetservice/repository/UserRepository.java
@@ -1,9 +1,9 @@
-package com.fse1.tweetservice.repository;
+package com.tweetapp.tweetservice.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.fse1.tweetservice.entity.UserEntity;
+import com.tweetapp.tweetservice.entity.UserEntity;
 
 @Repository
 public interface UserRepository extends MongoRepository<UserEntity, String> {
