+++ tweet-service/src/main/java/com/tweetapp/tweetservice/service/impl/UserServiceImpl.java
@@ -1,9 +1,9 @@
-package com.fse1.tweetservice.service.impl;
+package com.tweetapp.tweetservice.service.impl;
 
 import org.springframework.stereotype.Component;
 
-import com.fse1.tweetservice.dto.UserDto;
-import com.fse1.tweetservice.service.UserService;
+import com.tweetapp.tweetservice.dto.UserDto;
+import com.tweetapp.tweetservice.service.UserService;
 
 @Component
 public class UserServiceImpl implements UserService {
